requirements.txt
flask==3.0.3
gunicorn
uvicorn==0.30.1
openai==1.30.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==1.10.14
requests==2.32.3
asgiref==3.7.2
faiss-cpu==1.8.0
numpy==1.26.4
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
# Uncomment if you use login:
# authlib
# flask-login
